pyproject.toml
[project]
name = "stt-realtime-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=1.9.0",
    "googlemaps>=4.10.0",
    "numpy>=2.1.2",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.5.0",
    "websockets>=13.1",
]

[project.scripts]
stt-realtime-demo = "stt_realtime_demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"